requirements.txt
playwright
beautifulsoup4
requests
gspread
oauth2client
dnspython
lxml
apify-client
python-dotenv
haversine
